qt: localize activity dates and hand enum roles to QML as ints
Two small correctness riders on the activity model.

Transaction::dateTimeString formatted month names with QDateTime's
format strings, which in Qt 6 always render English regardless of the
user's locale, so every transaction date in the app was English-only.
Format through QLocale like the ban list dates already do.

The Status and Type roles returned raw C++ enum values, so QML
delegate comparisons against Transaction enum values worked only
through QVariant's enum coercion. Cast both to int at the model
boundary (data() and transactionDetails) to make the contract
explicit.

+++ qml/models/activitylistmodel.cpp
@@ -81,9 +81,9 @@ QVariant ActivityListModel::data(const QModelIndex &index, int role) const
     case LabelRole:
         return tx->label;
     case StatusRole:
-        return tx->status;
+        return static_cast<int>(tx->status);
     case TypeRole:
-        return tx->type;
+        return static_cast<int>(tx->type);
     case TxidRole:
         return tx->isPendingRequest ? QString{} : tx->txid;
     case CanBumpRole:
@@ -188,8 +188,8 @@ QVariantMap ActivityListModel::transactionDetails(const QSharedPointer<Transacti
         {"amount", tx->prettyAmount(m_display_unit)},
         {"date", tx->dateTimeString()},
         {"depth", tx->depth},
-        {"type", tx->type},
-        {"status", tx->status},
+        {"type", static_cast<int>(tx->type)},
+        {"status", static_cast<int>(tx->status)},
         {"countsForBalance", tx->countsForBalance},
         {"address", tx->address},
         {"label", tx->label},
